+++ burn-wgpu/src/kernel/mask/mod.rs
@@ -1,0 +1,5 @@
+mod base;
+mod mask_fill;
+mod mask_where;
+
+pub use base::*;
