+++ burn-wgpu/src/ops/base.rs
@@ -1,6 +1,6 @@
 use crate::{
     element::WgpuElement,
-    kernel::{self, cat, mask_fill, mask_fill_inplace, mask_where, mask_where_inplace},
+    kernel::{self, cat},
     pool::get_context,
     tensor::WgpuTensor,
     GraphicsApi, WgpuDevice,
@@ -82,33 +82,6 @@ impl<G: GraphicsApi> Init<G> {
         let tensor = kernel::into_continuous(tensor);
 
         WgpuTensor::new(tensor.context, shape, tensor.buffer)
-    }
-
-    pub fn mask_fill<E: WgpuElement, const D: usize>(
-        tensor: WgpuTensor<E, D>,
-        mask: WgpuTensor<u32, D>,
-        value: E,
-    ) -> WgpuTensor<E, D> {
-        if tensor.can_mut() {
-            return mask_fill_inplace(tensor, mask, value);
-        }
-
-        mask_fill(tensor, mask, value)
-    }
-
-    pub fn mask_where<E: WgpuElement, const D: usize>(
-        tensor: WgpuTensor<E, D>,
-        mask: WgpuTensor<u32, D>,
-        value: WgpuTensor<E, D>,
-    ) -> WgpuTensor<E, D> {
-        if tensor.can_mut_broadcast(&value) {
-            return mask_where_inplace(tensor, mask, value, false);
-        }
-        if value.can_mut_broadcast(&tensor) {
-            return mask_where_inplace(value, mask, tensor, true);
-        }
-
-        mask_where(tensor, mask, value)
     }
 
     pub fn cat<E: WgpuElement, const D: usize>(
